# raffleease-openapi.yaml
openapi: 3.1.0
info:
  title: Raffleease API
  description: |
    Raffleease is a RESTful API designed to support the management of online raffles for animal rescue associations. The platform follows a **multi-tenant architecture**, where most endpoints are scoped under a specific association ID to ensure data isolation across organizations.
    
    The API implements **JWT-based authentication** with secure refresh token handling via **HTTP-only, same-site, secure cookies**. **Role-based access control** governs permissions across different user types within each association.
    
    Search endpoints support **advanced filtering, sorting, and pagination** mechanisms to allow flexible querying of resources. All API responses follow a **consistent, standardized format**, and **error handling** is uniform across the platform to provide clarity for integrators.
    
    This documentation should clearly outline available endpoints, authentication flows, security requirements, and expected request and response formats to help developers integrate with the Raffleease platform effectively.    

    ## Key Features
    
    ### Multi-Tenant Architecture
    Most endpoints are scoped under `/associations/{associationId}` to ensure data isolation and support multiple organizations.
    
    ### JWT-Based Authentication
    The API uses JWT access tokens for authentication and secure HTTP-only cookies to manage refresh tokens.
    
    ### Role-Based Access Control
...
    
...
    
...
    
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /v1/associations/{associationId}/raffles/{raffleId}/tickets:
    get:
      tags:
        - Tickets
      summary: Search tickets
      description: |
        Searches for tickets in a raffle with filtering and pagination.
        
        **Filtering:**
        - Filter by ticket number
        - Filter by status (AVAILABLE, RESERVED, SOLD)
        - Filter by customer ID
        
        **Pagination:**
        - Page-based pagination with configurable page size
        - Returns total count and page metadata
      operationId: searchTickets
      parameters:
        - $ref: '#/components/parameters/associationId'
        - name: raffleId
          in: path
          required: true
          description: The unique identifier of the raffle
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 101
        - name: ticketNumber
          in: query
          required: false
          description: Filter by ticket number (exact match)
          schema:
            type: string
          example: "R1001"
        - name: status
          in: query
          required: false
          description: Filter by ticket status
          schema:
            $ref: '#/components/schemas/TicketStatus'
        - name: customerId
          in: query
          required: false
          description: Filter by customer ID (for sold tickets)
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 789
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/sizeParam'
        - $ref: '#/components/parameters/sortParam'
      responses:
        '200':
          description: Tickets retrieved successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/PageResponse'
                          - type: object
                            properties:
                              content:
                                type: array
                                items:
                                  $ref: '#/components/schemas/TicketResponse'
              example:
                success: true
                message: "Tickets retrieved successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data:
                  content:
                    - id: 789
                      ticketNumber: "R1001"
                      status: "AVAILABLE"
                      raffleId: 101
                  totalElements: 1000
                  totalPages: 50
                  size: 20
                  number: 0
                  first: true
                  last: false
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/associations/{associationId}/raffles/{raffleId}/tickets/random:
    get:
      tags:
        - Tickets
      summary: Get random tickets
      description: |
        Retrieves a specified number of random available tickets from a raffle.
        
        **Business Rules:**
        - Only returns tickets in AVAILABLE status
        - Randomization is truly random (not sequential)
        - Returns fewer tickets if not enough available
      operationId: getRandomTickets
      parameters:
        - $ref: '#/components/parameters/associationId'
        - name: raffleId
          in: path
          required: true
          description: The unique identifier of the raffle
          schema:
            type: integer
            format: int64
            minimum: 1
          example: 101
        - name: quantity
          in: query
          required: true
          description: Number of random tickets to retrieve
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 100
          example: 5
      responses:
        '200':
          description: Random tickets retrieved successfully
          headers:
            X-Rate-Limit-Remaining:
              $ref: '#/components/headers/X-Rate-Limit-Remaining'
            X-Rate-Limit-Reset:
              $ref: '#/components/headers/X-Rate-Limit-Reset'
            X-Rate-Limit-Limit:
              $ref: '#/components/headers/X-Rate-Limit-Limit'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TicketResponse'
              example:
                success: true
                message: "Random tickets retrieved successfully"
                timestamp: "2024-01-15T10:30:00Z"
                data:
                  - id: 789
                    ticketNumber: "R1001"
                    status: "AVAILABLE"
                    raffleId: 101
                  - id: 790
                    ticketNumber: "R1002"
                    status: "AVAILABLE"
                    raffleId: 101
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for API authentication
  parameters:
    associationId:
      name: associationId
      in: path
      required: true
      description: The unique identifier of the association for multi-tenant resource access
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 123
    pageParam:
      name: page
      in: query
      required: false
      description: Zero-based page number for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0
    sizeParam:
      name: size
      in: query
      required: false
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    sortParam:
      name: sort
      in: query
      required: false
      description: |
        Sorting criteria in the format: property,direction. 
        Direction can be 'asc' or 'desc'. 
        Multiple sort parameters are supported.
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
      examples:
        singleSort:
          value: ["createdAt,desc"]
          description: Sort by creation date descending
        multipleSort:
          value: ["status,asc", "createdAt,desc"]
          description: Sort by status ascending, then by creation date descending
  headers:
    X-Rate-Limit-Remaining:
      description: Number of requests remaining in current rate limit window
      schema:
        type: integer
    X-Rate-Limit-Reset:
      description: Time when rate limit window resets (Unix timestamp)
      schema:
        type: integer
    X-Rate-Limit-Limit:
      description: Total number of requests allowed in rate limit window
      schema:
        type: integer
  schemas:
...
    ApiResponse:
      type: object
      discriminator:
        propertyName: success
        mapping:
          true: '#/components/schemas/SuccessResponse'
          false: '#/components/schemas/ErrorResponse'
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
          description: Indicates whether the request was successful
        message:
          type: string
          description: Human-readable message describing the result
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the response was generated
      example:
        success: true
        message: "Operation completed successfully"
        timestamp: "2024-01-15T10:30:00Z"
    SuccessResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              enum: [true]
            data:
              description: The response payload containing the requested data
              oneOf:
                - type: object
                - type: array
                - type: string
                - type: number
                - type: boolean
                - type: "null"
      example:
        success: true
        message: "Data retrieved successfully"
        timestamp: "2024-01-15T10:30:00Z"
        data:
          id: 123
          name: "Example"
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - statusCode
            - statusText
            - code
          properties:
            success:
              type: boolean
              enum: [false]
            statusCode:
              type: integer
              description: HTTP status code
              minimum: 400
              maximum: 599
            statusText:
              type: string
              description: HTTP status text
            code:
              type: string
              description: Application-specific error code for programmatic handling
              enum:
                - NOT_FOUND
                - BAD_REQUEST
                - UNAUTHORIZED
                - ACCESS_DENIED
                - CONFLICT
                - FILE_STORAGE_ERROR
                - USER_NOT_FOUND
                - EMAIL_VERIFICATION_FAILED
                - PASSWORD_RESET_FAILED
                - MAIL_ERROR
                - INVALID_REQUEST
                - MISSING_PARAMETER
                - BUSINESS_ERROR
                - UNEXPECTED_ERROR
                - VALIDATION_ERROR
                - VALUE_ALREADY_EXISTS
                - RATE_LIMIT_EXCEEDED
                - IMAGE_LIMIT_EXCEEDED
                - EMAIL_SAME_AS_CURRENT
                - EMAIL_UPDATE_TOKEN_EXPIRED
                - EMAIL_UPDATE_TOKEN_INVALID
                - EMAIL_NO_LONGER_AVAILABLE
                - CURRENT_PASSWORD_INCORRECT
                - PASSWORD_SAME_AS_CURRENT
                - ROLE_UPDATE_SELF_DENIED
                - ROLE_UPDATE_ADMIN_DENIED
                - ADMIN_DISABLE_SELF_DENIED
                - ADMIN_CREATE_ADMIN_DENIED
                - INSUFFICIENT_TICKETS_AVAILABLE
      example:
        success: false
        message: "Resource not found"
        timestamp: "2024-01-15T10:30:00Z"
        statusCode: 404
        statusText: "Not Found"
        code: "NOT_FOUND"
...
    PageResponse:
      type: object
      required:
        - content
        - totalElements
        - totalPages
        - size
        - number
        - first
        - last
        - numberOfElements
      properties:
        content:
          type: array
          items: {}
          description: Array of items for the current page
        totalElements:
          type: integer
          format: int64
          description: Total number of elements across all pages
        totalPages:
          type: integer
          description: Total number of pages
        size:
          type: integer
          description: Number of items per page
        number:
          type: integer
          description: Current page number (zero-based)
        first:
          type: boolean
          description: Whether this is the first page
        last:
          type: boolean
          description: Whether this is the last page
        numberOfElements:
          type: integer
          description: Number of elements in the current page
        empty:
          type: boolean
          description: Whether the page is empty
      example:
        content: []
        totalElements: 150
        totalPages: 8
        size: 20
        number: 0
        first: true
        last: false
        numberOfElements: 20
        empty: false

    # Common Enums
...
    TicketStatus:
      type: string
      enum:
        - AVAILABLE
        - RESERVED
        - SOLD
      description: |
        The status of a ticket:
        - **AVAILABLE**: Ticket can be purchased
        - **RESERVED**: Ticket is temporarily held in a cart
        - **SOLD**: Ticket has been purchased
...
    TicketResponse:
      type: object
      required:
        - id
        - ticketNumber
        - status
        - raffleId
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier of the ticket
          example: 789
        ticketNumber:
          type: string
          description: The unique ticket number within the raffle
          example: "R1001"
        status:
          $ref: '#/components/schemas/TicketStatus'
        raffleId:
          type: integer
          format: int64
          description: The ID of the raffle this ticket belongs to
          example: 101
        customerId:
          type: integer
          format: int64
          description: The ID of the customer who purchased the ticket (if sold)
          example: 456
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: When the ticket was created
          example: "2024-01-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          description: When the ticket was last updated
          example: "2024-01-15T10:30:00Z"
...
  responses:
...
    Unauthorized:
      description: Unauthorized - Authentication required or token invalid
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Authentication required"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 401
            statusText: "Unauthorized"
            code: "UNAUTHORIZED"
    Forbidden:
      description: Forbidden - Insufficient permissions to access the resource
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Access denied - insufficient permissions"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 403
            statusText: "Forbidden"
            code: "ACCESS_DENIED"
    NotFound:
      description: Not Found - The requested resource was not found
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Resource not found"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 404
            statusText: "Not Found"
            code: "NOT_FOUND"
...
    TooManyRequests:
      description: Too Many Requests - Rate limit exceeded
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
        Retry-After:
          description: Number of seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "Rate limit exceeded. Please try again later."
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 429
            statusText: "Too Many Requests"
            code: "RATE_LIMIT_EXCEEDED"
    InternalServerError:
      description: Internal Server Error - An unexpected error occurred
      headers:
        X-Rate-Limit-Remaining:
          $ref: '#/components/headers/X-Rate-Limit-Remaining'
        X-Rate-Limit-Reset:
          $ref: '#/components/headers/X-Rate-Limit-Reset'
        X-Rate-Limit-Limit:
          $ref: '#/components/headers/X-Rate-Limit-Limit'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "An unexpected error occurred"
            timestamp: "2024-01-15T10:30:00Z"
            statusCode: 500
            statusText: "Internal Server Error"
            code: "UNEXPECTED_ERROR" 